Move drag handle to it's own file
